refactor: google analytics를 custom hook으로 추상화
