+++ src/pages/_document.tsx
@@ -29,8 +29,8 @@ const Document = () => {
 
         {/* Google Analytics */}
         <Script
-          strategy='afterInteractive'
           src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID}`}
+          strategy='afterInteractive'
         />
       </body>
     </Html>
