+++ src/utils/helpers/gtag.ts
@@ -1,8 +1,6 @@
 // 조회수 측정
 export const pageview = (url: URL) => {
-  if (typeof window === undefined) return;
-
-  window.gtag('config', process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID, {
+  gtag('config', process.env.NEXT_PUBLIC_GOOGLE_ANALYTICS_ID, {
     page_path: url,
   });
 };
@@ -19,9 +17,7 @@ export const event = ({
   label: string;
   value: number;
 }) => {
-  if (typeof window === undefined) return;
-
-  window.gtag('event', action, {
+  gtag('event', action, {
     event_category: category,
     event_label: label,
     value,
